+++ docs/changelog.md
@@ -6,12 +6,14 @@ import GH from './GH.vue';
 
 ## unreleased
 
-- Add [`default:update`](/reference/default.md#default-update-yes-no) <GH issue="171" pr="175"/>
-- Add [`context`](./reference/context.md). See [Guide: Pass context to
-  functions](./guide/context.md) <GH issue="68" pr="176"/>
+## v1.7.0
+
 - Remove default value of [`arg:context:regex`](./reference/arg.md) <GH issue="68" pr="176"/>
   - To get v1.6.0 behavior, configure: `goverter:arg:context:regex ^ctx|^context`
   - The recommended way to configure context is to use [`context`](./reference/context.md).
+- Add [`context`](./reference/context.md). See [Guide: Pass context to
+  functions](./guide/context.md) <GH issue="68" pr="176"/>
+- Add [`default:update`](/reference/default.md#default-update-yes-no) <GH issue="171" pr="175"/>
 - Fix not setting `nil` on map when value is `nil`. <GH issue="173" pr="175"/>
 
 ## v1.6.0